Create feature flag for optional workpass validation screen
The previous rollout did not consider the work pass validation screen and the
actual user validation as two separate features. As such, users could bypass the
validation on the frontend but be unable to make calls because they were blocked
in the backend.

This creates a separate feature flag for controlling the display of the screen,
which will be replaced by a different flag indicating mandatory/optional
workpass validation.