Create build config for shared package
The shared package spontaneously stopped working for me in development,
plausibly due to the typescript files not being compiled in the version of
node/npm I am using (15.14.0, 7.7.6).

This change sets up a build configuration for the shared package so that the js
files can be used instead of ts files. This probably slows down development
because the shared modules must be constantly rebuilt, but it works.