Add user experiment config and Twilio edge experiment (#50)
* Add table for per-user experiments

This table will be queried by the relevant services in order to determine what
flags to enable for the user. The most pertinent use of this is to tweak call
settings on a per-user basis.

* Add UserExperimentConfig service

This module handles two major responsibilities:
- Generating a new UserExperimentConfig with all of the flags that we want via
  RNG.
- CRUD for UserExperimentConfig. Similarly to validations, a new row is created
  when an old one is not found. If an old one is found but is missing some
  expected experiments, it is updated accordingly.

* Create build config for shared package

The shared package spontaneously stopped working for me in development,
plausibly due to the typescript files not being compiled in the version of
node/npm I am using (15.14.0, 7.7.6).

This change sets up a build configuration for the shared package so that the js
files can be used instead of ts files. This probably slows down development
because the shared modules must be constantly rebuilt, but it works.

* Add edge experiment to frontend

Co-authored-by: Glen <glen@Glens-MacBook-Pro.local>