fix: use local .npm directory for npm installation
This fixes npm creating a symlink in the global node_modules directory when
performing `npm link`, which may not be possible if the global node_modules
directory is read-only.

This also enables caching of npm packages in a local directory, which will be
useful for CI caching down the road.