chore: create super app dir structure
The eventual goal is to have something like:

.
├── apps/
│   ├── common
│   └── super-app
│       ├── pro-bono
│       ├── personal-calls
│       └── twc2
├── services/
│   ├── superset
│   ├── strapi
│   └── personal-calls
└── infra

This should make it relatively simple to:

- deploy a single frontend app
- develop sub-apps in the super app while sharing code
- develop services in isolation
- make changes to cloud infrastructure

We want to split the personal call code into app + service, but that's going to
be a big effort. Keep it in the top-level personal-calls/ in the meantime.

If we need to share frontend code with the personal calls app, some options are:

- Duplicate the code
- Create a local npm module for the shared code