super-app: add eslint and prettier
Add linting and formatting scripts:

- `npm run lint` checks for JS/TS linting errors
- `npm run format` formats JS/TS code
- `npm run format:check` checks JS/TS code for formatting errors.

and corresponding `make` rules that can be run locally, and format the
code accordingly. When run in CI, the `npm run` form is used, which
doesn't require that `make` be present in the CI runner. The CI checks
would preferably run in parallel instead of failing fast, but we'll
figure that out later.

Using CI is more robust than the common pattern of using `lint-staged` +
`husky` and checking using a Git precommit hook (especially good since
at least one contributor doesn't create commits with git, aka me). Also,
precommit hooks are quite clunky in a monorepo setup like this (would we
want to run precommits for _all_ projects?) and can be quite slow.