personal-calls: create shareable directories
Move code that we want to share with other apps into a common/
directory. This code won't _actually_ be split off into a common npm
package, because that would pin us to the old versions of Material UI
and React used by the personal calls frontend.

Instead, just leave some notes for ourselves so that we can manage this
bit of duplication in a sensible manner.