super-app: add themed button PoC
Copy the "common" code from `personal-calls/`, modify it for
@mui/material, and show that it works by using the themes.

Material UI v4 -> v5 is an unfortunately massive change. This is
certainly not the best way to achieve the theming result, but it's the
best I could come up with at the moment, and we're too early in
development for this to block our progress. We can clean this up later.