build: add terraform infra for staging
This commit adds terraform infra setups for staging and some basic
Makefile scripts for deployment.

The staging GCP project `call-home-staging` and its GCS state bucket
`call-home-staging-terraform-state-bucket` should be managed outside of
this terraform project, and they have already been created manually. The
state bucket however should ideally be hosted on a separate project
`call-home-infra`, and moving this bucket over may require a name change,
which should be addressed in the future if necessary.

This commit also assumes that staging and production infra setups are to
be hosted separately under independent directories (`staging/` vs.
`production/`). Modules or workspaces should be introduced in the future
to promote reuse and keep the configuration environment-agnostic.