super-app: add smoke test from frontend
This queries the CMS from the frontend to see if it is running. This
probably won't be permanent because in production, we'll query the CMS
for nearly every page, so we'd get the same information.

Unfortunately, this hardcodes 1337 as the proxy port, which effectively
means that for the test to work, Strapi has to be run on 1337. This
_should_ be fixable, but we can do that later.