super-app: add react-router
Add a simple `react-router-dom` setup by moving the testing content from
`App.tsx` to a `/scratch` route. We can reuse this route for whatever
other content we want to test out.

This uses the old component-based routing instead of the react-router v6
`createXRouter` APIs. This is largely because this makes it easier to
use a different router in testing vs in production (the "browser router"
isn't available in testing; the "memory router" is recommended instead).
The <App/> component is wrapped by the <(Browser|Memory)Router/>
component, and we can easily change routers by specifying the right
wrapping router in the right environment.

I'm not sure if having the router outside of the app is actually a good
pattern (react-router seems to recommend this) or more of a hack (we now
have some providers inside the <App/> and some outside of it, which
feels inconsistent), but we can revisit this pattern later.