super-app: create visual outline of support detail page
This creates the rough outline of the TWC2 support detail page (minus
the Facebook links). This is mostly visual and the content is hardcoded
placeholders for now.

This is intentionally TWC2-only. We'll worry about Pro Bono SG when the
time comes, but it shouldn't be too hard to generalise.