Add service card component
use MUI Card to create the different services' reusable cards. These cards should be able to take it the values from a separate service. also modified Scratch to Home Page and add the card components to Home Page. Include routing inside this commit as well. Lastly, refactored some of the services and interfaces for better reusability and extendability.