super-app: add navbar to LandingPage
Also, simplify the NavBar styles a bit. This makes the NavBar visually
correct, but the navigation doesn't work yet. We'll fix that in the next
commit.

The LandingPage component has padding on the Container. We don't want
the NavBar to be padded, but we do want it vertically centered (aka
inside the container).

The solution is to split the container into the NavBar, and the "above
the NavBar" portion, and apply the padding to only the stuff above the
NavBar. This also lets us constrain its height so that it doesn't eat
into the NavBar's space.