super-app: sync directories when deploying
We are using `gsutil cp -r`, which copies over all files to GCS.
However, each frontend build may emit a different set of filenames,
(e.g. the JS is hashed to produce a name like main.88af4d.js), so this
causes build artifacts from different builds to accumulate in the bucket.

Use `gsutil rsync` instead, which syncs the states of the directories,
which is semantically what we want (we want the GCS bucket to contain
the latest build and _only_ the latest build).