apps/common: split out common code from personal-calls
Create a 'shared' module for our various apps by splitting it out from
`personal-calls/frontend`. It will be 'shared' by copying it to the
various locations. Having multiple copies is not ideal, but few
alternatives exist, and I've documented them in `apps/common/README.md`.

This isn't quite ready to be copied to `apps/super-app` yet, but we'll
clean those up in later commits.