apps-common: unify Theme.ts
The only major difference between the `Theme.ts` copies is that
`apps/super-app` defines a bunch of Typography variants, including h5.
`personal-calls/frontend` uses variant=h5 as a shorthand to get the
styles we want, so we need to replace that with something else. All of
the other Typography changes are no-ops because they aren't used.

In nearly all places, we've been using the h5 variant for either h1 or
h2, so it's pretty clear that this just about as big as we want them to
be. Define a new h2 variant that looks exactly the same as @mui's
default h5 (values taken from
https://mui.com/material-ui/customization/default-theme/) and use that
instead.